Stream sample counting

+++ source/TS.NET.Tests/Triggers/Situations/AnyEdgeTriggerSituationsI16.cs
@@ -14,7 +14,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(short.MaxValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -33,7 +33,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(short.MinValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -52,7 +52,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(short.MaxValue);
         situation.Input.Span.Slice(300, 100).Fill(short.MaxValue);
@@ -72,7 +72,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(short.MinValue);
         situation.Input.Span.Slice(300, 100).Fill(short.MinValue);
@@ -92,7 +92,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[2];
+        situation.ExpectedWindowEndIndices = new ulong[2];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(short.MaxValue);
         situation.Input.Span.Slice(10200, 100).Fill(short.MaxValue);
@@ -113,7 +113,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[2];
+        situation.ExpectedWindowEndIndices = new ulong[2];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(short.MinValue);
         situation.Input.Span.Slice(10200, 100).Fill(short.MinValue);
@@ -134,7 +134,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(-12800);
         situation.Input.Span.Slice(0, 100).Fill(short.MaxValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -153,7 +153,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(12800);
         situation.Input.Span.Slice(0, 100).Fill(short.MinValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -172,7 +172,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span[100] = short.MaxValue;
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -191,7 +191,7 @@ internal class AnyEdgeTriggerSituationsI16
             ChunkCount = 1
         };
         situation.Input = new short[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span[100] = short.MinValue;
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
