+++ source/TS.NET.Tests/Triggers/Situations/AnyEdgeTriggerSituationsI8.cs
@@ -14,7 +14,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(sbyte.MaxValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -33,7 +33,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(sbyte.MinValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -52,7 +52,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(sbyte.MaxValue);
         situation.Input.Span.Slice(300, 100).Fill(sbyte.MaxValue);
@@ -72,7 +72,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(sbyte.MinValue);
         situation.Input.Span.Slice(300, 100).Fill(sbyte.MinValue);
@@ -92,7 +92,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[2];
+        situation.ExpectedWindowEndIndices = new ulong[2];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(sbyte.MaxValue);
         situation.Input.Span.Slice(10200, 100).Fill(sbyte.MaxValue);
@@ -113,7 +113,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[2];
+        situation.ExpectedWindowEndIndices = new ulong[2];
         situation.Input.Span.Fill(0);
         situation.Input.Span.Slice(100, 100).Fill(sbyte.MinValue);
         situation.Input.Span.Slice(10200, 100).Fill(sbyte.MinValue);
@@ -134,7 +134,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(-50);
         situation.Input.Span.Slice(0, 100).Fill(sbyte.MaxValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -153,7 +153,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(50);
         situation.Input.Span.Slice(0, 100).Fill(sbyte.MinValue);
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -172,7 +172,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span[100] = sbyte.MaxValue;
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
@@ -191,7 +191,7 @@ internal class AnyEdgeTriggerSituationsI8
             ChunkCount = 1
         };
         situation.Input = new sbyte[situation.ChunkSize * situation.ChunkCount];
-        situation.ExpectedWindowEndIndices = new int[1];
+        situation.ExpectedWindowEndIndices = new ulong[1];
         situation.Input.Span.Fill(0);
         situation.Input.Span[100] = sbyte.MinValue;
         situation.ExpectedWindowEndIndices.Span[0] = 10100;
